refactor: userInfo, Auth 라우터 위치 변경
