+++ backend/app.js
@@ -4,10 +4,10 @@ var path = require("path");
 var cookieParser = require("cookie-parser");
 var logger = require("morgan");
 
-var indexRouter = require("./routes/index");
-var usersRouter = require("./routes/users");
 const apiRouter = require("./routes/api");
-const authRouter = require("./routes/api/auth");
+const authRouter = require("./routes/auth");
+const infoRouter = require("./routes/user");
+
 const cors = require("cors");
 
 var app = express();
@@ -23,10 +23,9 @@ app.use(cookieParser());
 app.use(express.static(path.join(__dirname, "public")));
 
 app.use(cors());
-app.use("/", indexRouter);
-app.use("/users", usersRouter);
 app.use("/api", apiRouter);
 app.use("/auth", authRouter);
+app.use("/info", infoRouter);
 
 //mongodb연결을위해 비밀번호 가림
 const dotenv = require("dotenv");
